+++ agent/src/errors.rs
@@ -1,6 +1,10 @@
+use log::SetLoggerError;
 use thiserror::Error;
 
-use crate::db::database::DatabaseSaveError;
+use crate::{
+    db::database::{DatabaseOpenError, DatabaseSaveError},
+    file_client::errors::RCloneCommandError,
+};
 
 #[derive(Debug, Error)]
 pub enum AuthError {
@@ -48,6 +52,8 @@ pub enum ClientRuntimeError {
     NetRuntimeError(NetRunError),
     #[error("Network initialisation error: {0}")]
     NetInitError(NetInitError),
+    #[error("File sync error (rclone): {0}")]
+    FileSyncError(RCloneCommandError),
 }
 
 impl ClientRuntimeError {
@@ -62,4 +68,34 @@ impl ClientRuntimeError {
     pub fn net_init(error: NetInitError) -> Self {
         return Self::NetInitError(error);
     }
+}
+
+#[derive(Debug, Error)]
+pub enum ConfigError {
+    #[error("Failed to open ini file: {0}")]
+    FailedToOpenIniFile(String),
+    #[error("Missing required section: {0} in ini file")]
+    MissingSection(String),
+    #[error("Missing required key: {0} in section: {1} of ini file")]
+    MissingKey(String, String),
+}
+
+#[derive(Debug, Error)]
+pub enum RequirementMissing {
+    #[error("RClone is not installed or not found in PATH")]
+    RCloneNotFound,
+}
+
+#[derive(Debug, Error)]
+pub enum ClientInitError {
+    #[error("Configuration error: {0}")]
+    ConfigError(ConfigError),
+    #[error("Failed to open database: {0} (path: {1}")]
+    OpenDbError(DatabaseOpenError, String),
+    #[error("Failed to create logger (simple_logger): {0}")]
+    LogInitError(SetLoggerError),
+    #[error("Invalid log level")]
+    LogLevelError,
+    #[error("Requirement missing: {0}")]
+    RequirementMissing(RequirementMissing),
 }