+++ server/crates/application/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod agent;
 pub mod database;
+pub mod file_server;
 pub mod net;