+++ server/crates/infrastructure/src/file_server/rclone/errors.rs
@@ -1,0 +1,9 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum FileServerErrors {
+    #[error("Reached timeout while waiting for the file server to start")]
+    TimeoutError,
+    #[error("The file server proxy is already started")]
+    ProxyAlreadyStarted,
+}