+++ server/crates/infrastructure/src/net/rustls/mod.rs
@@ -1,4 +1,2 @@
-pub mod certgen;
 pub mod connection;
-pub mod errors;
 pub mod network_manager;